% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the Fars data on a US stat map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number representing a US state.}

\item{year}{The year the data was obtained.}
}
\value{
Returns a map with the Fars data of the given year
plotted to a given US states map.
}
\description{
This function takes one or more year and a US
state.num as input and plots the Fars data for the
given years on the states map. If a year or the state number
does not exist in the data the function will return an error.
}
\examples{
\dontrun{
fars_map_state(45,2013)
}

}
