% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the csv file}
}
\value{
A tibble generated from the csv file
}
\description{
This function reads a csv file and returns a tibble. The function
returns an Error if the file cannot be found.
}
