% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Fars data for given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector of years.}
}
\value{
Returns a List of tibbles with MONTH and year of each
given year
}
\description{
This function takes one or more years as an input
and returns a list of tibbles which for each year lists
the MONTH and year.
Returns an error if data for a given year is not found.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
