% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year the data was obtained}
}
\value{
A string in the format 'accident_year.csv.bz2'
that can be used as filename
}
\description{
This function takes a year as numeric input and returns a filname
in the format 'accident_year.csv.bz2'.
}
\examples{
\dontrun{
makefilename(2014)
}

}
